rename to cioos package
