+++ README.md
@@ -1,23 +1,23 @@
 # CKAN Metadata Format Generator
 
 This project provides a tool to generate various metadata formats and standards
-from the Hakai(CIOOS) metadata catalog form `yaml` format. It aims to
+from the CIOOS metadata catalog form `yaml` format. It aims to
 facilitate the sharing and citation of datasets by converting metadata records into
 widely recognized standards.
 
 ## Currently Supported Standards
 
 - `citation.cff`: Citation File Format, a human and machine-readable file format
-which provides citation metadata for software.
+  which provides citation metadata for software.
 
 ## Features
 
 - **Easy Integration**: Seamlessly integrates with existing CKAN instances to
-fetch catalog records.
+  fetch catalog records.
 - **Extensible**: Designed with extensibility in mind, allowing for the addition
-of more metadata standards in the future.
+  of more metadata standards in the future.
 - **User-Friendly**: Provides a simple and intuitive interface for generating
-metadata files.
+  metadata files.
 
 ## Getting Started
 
@@ -32,39 +32,38 @@ To use this tool, follow these steps:
 
 2. **Install Dependencies:**
 
-    ```bash
-    pip install -e .
-    ```
+   ```bash
+   pip install -e .
+   ```
 
 3. **Run the Tool:**
 
-    ```bash
-    python hakai_metadata_conversion 
-    ```
+   ```bash
+   python cioos_metadata_conversion
+   ```
 
-    For more information
+   For more information
 
-    ```bash
-    python hakai_metadata_conversion --help
-    ```
+   ```bash
+   python cioos_metadata_conversion --help
+   ```
 
 ## Use within an action
 
 The tool can be use within a github action by adding the following step within your action:
 
 ```yaml
-job: 
-    sync:
-        step:
-            - use: action/checkout@v4
-            - name: Sync metadata
-              use: hakaiinstitute/hakai-metadata-conversion
-              with:
-                - input: url or file path within repo
-                - output-file: CITATION.cff
-                - output-format: cff
+job:
+  sync:
+    step:
+      - use: action/checkout@v4
+      - name: Sync metadata
+        use: hakaiinstitute/hakai-metadata-conversion
+        with:
+          - input: url or file path within repo
+          - output-file: CITATION.cff
+          - output-format: cff
 ```
-
 
 ## How to Contribute
